app/Collections/requirements.txt
mysql-connector-python==8.0.25
requests==2.28.2
python-slugify==5.0.2